Bootstrap evaluation using LLVM code generation
Bootstrap evaluation using llvm code generation

LLVM code generation is done using a mix of :
- glue IR code that loops over the vector, generates function
  calls 
- byte-code files generated from simple c++ functions using 
  clang .
The glue-code and pre-compiled byte code are merged and 
optimized together.

Expressions are specified using a "tree builder" where each 
node is an arrow vector, or a binary/unary function.

During code generation, the expressions are "decomposed" so 
that the value array and bitmap array are evaluated separately 
to compute the expression result. This avoids the use of too 
many branch/conditional instructions , and
hence, can be vectorized efficiently.

Support added for arithmetic and logical expressions on 
numeric types.

Travis CI support added for build on ubuntu.