[GDV]Make use of the modular features of cmake
Separate out the public and private target dependencies.

For arrow, export an interface target. This avoids the need to add
include dirs for each dependency on arrow.

Removed dependency on gtest. Instead, build it as an external project.
This is the recommended practice for googletest.

For pre-compiled files, generate the bitcode files for each of them
independently and then, link them to generate a unified bitcode file.
Removed cpplint exceptions since there is no more sourcing of .cc
files.

Separate out the public include files from private includes, and 
add them in the dependency list in cmake.

pass the bytecode filepath from cmake